// src/domain/models/auth/Permission.ts
export type Permission =
  | "STUDENTS_WRITE"
  | "STUDENTS_READ"
  | "COORDINATORS_WRITE"
  | "COORDINATORS_READ"
  | "TEACHERS_WRITE"
  | "TEACHERS_READ"
  | "PROFILES_WRITE"
  | "SUBJECTS_WRITE"
  | "SUBJECTS_READ"
  | "PROGRAMS_WRITE"
  | "PROGRAMS_READ"
  | "GRADES_WRITE"
  | "GRADES_READ"
  | "ATTENDANCE_WRITE"
  | "ATTENDANCE_READ"
  | "ABSENCE_JUSTIFICATIONS_WRITE"
  | "ABSENCE_JUSTIFICATIONS_READ"
  | "ATTENDANCE_OBSERVATIONS_WRITE"
  | "ATTENDANCE_OBSERVATIONS_READ"
  | "STUDENT_OBSERVATIONS_WRITE"
  | "STUDENT_OBSERVATIONS_READ"
  | "ALERT_OBSERVATIONS_WRITE"
  | "ALERT_OBSERVATIONS_READ"
  | "REPORTS_READ"
  | "EXCEL_IO_WRITE"
  | "GROUPS_WRITE"
  | "GROUPS_READ"
  | "CONFIGURATION_WRITE"
  | "CONFIGURATION_READ";